cleanup grades #33 - & start cleanup old dir
