+++ src/app/Conponents/grades/grade-detail/grade-detail.component.html
@@ -28,6 +28,7 @@
     </div>
     <div>
         <span class="data-header">{{"grades.note" | translate}}</span>
+<!--        TODO: this should not be the mark but the note that the teacher added to the grade (api-issue #36)-->
         <span class="data-value">{{grade?.mark}}</span>
     </div>
 </div>
