--- src/app/Conponents/grades/parent-child-preview/parent-child-preview.component.scss
@@ -1,30 +1,0 @@
-@import 'src/styles.scss';
-
-.parent-child-preview-container{
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    
-    margin: 20px 0;
-    gap: 20px;
-
-    color: var(--teachu-primary);
-
-    .grade-container{
-        display: flex;
-        flex-direction: row;
-        flex-wrap: nowrap;
-        gap: 10px;
-    }
-}
-
-@media (max-width: 600px) {
-    .grade-container{
-        grid-column: 1 / span 2;
-    }
-
-    .parent-child-preview-container{
-        grid-template-columns: 50% 50%;
-        grid-template-rows: 60px 60px;
-    }
-}
