--- src/app/DTOs/xx_old/ChildDTO.ts
@@ -1,5 +1,0 @@
-export interface ChildDTO {
-    id: number;
-    firstName: string;
-    lastName: string;
-}
