--- src/app/Pages/grades/grades-parent-view/grades-parent-view.component.html
@@ -1,10 +1,0 @@
-<ng-container *ngIf="loaded">
-    <mat-tab-group>
-        <mat-tab *ngFor="let childMarks of childrenMarks"
-                 label="{{childMarks.firstName}}">
-            <app-parent-child-preview [childData]="childMarks"></app-parent-child-preview>
-            <app-grades-student-view [grades]="getChildSemesters(childMarks)"></app-grades-student-view>
-        </mat-tab>
-    </mat-tab-group>
-</ng-container>
-
