--- src/app/Pages/grades/grades-parent-view/grades-parent-view.component.scss
@@ -1,4 +1,0 @@
-:host ::ng-deep .mat-tab-body-content {
-    height: max-content !important;
-    overflow: hidden !important;
-}
