+++ src/app/Pages/grades/grades.component.html
@@ -1,1 +1,18 @@
-<app-grades-student-view [grades]="gradeSemesters"></app-grades-student-view>
+<div class="student-view-container">
+    <app-recent-exams [amount]="RECENT_EXAMS_AMOUNT" [semesterData]="allSemesterDTOs" (examClickedEvent)="recentExamClicked($event)"></app-recent-exams>
+    <div [class]="getGridClass()">
+        <app-menu-tree [style.display]="getMenuTreeDisplay()"
+                       (leaveClickedEvent)="menuTreeLeaveClicked($event)"
+                       [treeData]="examMenuTree">
+        </app-menu-tree>
+        <app-class-grades [style.display]="getClassGradesDisplay()"
+                          [selectedSubject]="selectedSubject"
+                          (gradeSelected)="gradeSelected($event)"
+                          (openMenuClicked)="openMenu()">
+        </app-class-grades>
+        <app-grade-detail [style.display]="getGradesDisplay()"
+                          [grade]="selectedGrade"
+                          (closeGradeDetail)="closeGradeDetails()">
+        </app-grade-detail>
+    </div>
+</div>
