+++ src/app/Pages/login/login.component.ts
@@ -3,8 +3,8 @@ import { FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { appRoutes } from 'src/app/Config/appRoutes';
 import { endpoints } from 'src/app/Config/endpoints';
-import { LoginDTO } from 'src/app/DTOs/xx_old/LoginDTO';
-import { TokenDTO } from 'src/app/DTOs/xx_old/TokenDTO';
+import { LoginDTO } from 'src/app/DTOs/Authentication/LoginDTO';
+import { TokenDTO } from 'src/app/DTOs/Authentication/TokenDTO';
 import { ApiService } from 'src/app/Framework/API/api.service';
 import { ErrorHandlingService } from 'src/app/Framework/API/error-handling.service';
 import { TokenService } from 'src/app/Framework/API/token.service';
