+++ src/app/app.module.ts
@@ -42,13 +42,10 @@ import { GroupChatDialogComponent } from './Pages/chat/group-chat-dialog/group-c
 import { PrivateChatDialogComponent } from './Pages/chat/private-chat-dialog/private-chat-dialog.component';
 import { AbsencesComponent } from './Pages/absences/absences.component';
 import { SchoolInfosComponent } from './Pages/school-infos/school-infos.component';
-import {GradesStudentViewComponent} from "./Pages/grades/grades-student-view/grades-student-view.component";
-import { GradesParentViewComponent } from './Pages/grades/grades-parent-view/grades-parent-view.component';
 import { RecentExamsComponent } from './Conponents/grades/recent-exams/recent-exams.component';
 import { ClassGradesComponent } from './Conponents/grades/class-grades/class-grades.component';
 import { GradeDetailComponent } from './Conponents/grades/grade-detail/grade-detail.component';
 import { LessonDetailsComponent } from './Pages/timetable/lesson-details/lesson-details.component';
-import { ParentChildPreviewComponent } from './Conponents/grades/parent-child-preview/parent-child-preview.component';
 import { ClasslistComponent } from './Pages/classlist/classlist.component';
 import { ClasslistListComponent } from './Pages/classlist/classlist-list/classlist-list.component';
 import { ClasslistDetailComponent } from './Pages/classlist/classlist-detail/classlist-detail.component';
@@ -85,8 +82,6 @@ export function HttpLoaderFactory(http: HttpClient) {
         GradesComponent,
         ReadonlyFieldComponent,
         ChangePasswordDialogComponent,
-        GradesStudentViewComponent,
-        GradesParentViewComponent,
         ChatOverviewComponent,
         ChatConversationComponent,
         ChatConversationInfoComponent,
@@ -98,8 +93,6 @@ export function HttpLoaderFactory(http: HttpClient) {
         ClassGradesComponent,
         GradeDetailComponent,
         LessonDetailsComponent,
-        ParentChildPreviewComponent,
-        ParentChildPreviewComponent,
         ClasslistComponent,
         ClasslistListComponent,
         ClasslistDetailComponent,
